requirements.txt
Flask==2.0.3
gunicorn==20.1.0
pycryptodomex==3.16.0
